Introduced core module with TaskioGroup.
The TaskioGroup will check all commands from the sources defined
at the configuration and add to the first level of commands, groups
will be added also. The sources primarily are modules present in
the pythonpath of the program being executed.

Becase the children of those groups will be located at the same
source, they will be added to the root, for this reason we remove
them after checking all commands in the source.

We still need check if multi level groups will work fine, at this
time we are just trying to solve the first level.

Refs: #5