Finished basic idea to strucutre an taskio application.
There will be a TaskioMultiCommand that will be the root of this
application, a decorator was added to provide a better experience.

We need taskio groups alos to keep propagating the taskio custom
look and feel.

Basically we keep doing what cli is doing but with some changes in
the initialization. We need a config structure in the beginning to
provide something to resolve the first level commands.

The rest of the operation is basically using our decorators in the
same way click does.

Refs: #2