Perfomed context refactory.
Renamed TaskioTestContext to TaskioContext and the TaskioContext
to TaskioCliContext.

Now the TaskioContext is the one used in the execution wiht the
muilt command and groups from the taskio and the TaskioCliContext
is the base to create context objects to be referenced by the
click.pass_context function.

The TaskioCliContext will have a reference to the TaskioContext
making it easy to access the loader used to resolve commands.

Refs: #17