refactor(core): renamed TaskioMultiCommand to TaskioRootGroup
Click will depreciate MultiCommand at the next release so we
decided to replace the use of Multicommand by Group.

As Group is descendant of MultiCommand, it was just a matter of
replace the class being extended. The TaskioMultiCommand class
was renamed to TaskioRootGroup for naming consistency.

With this commit the root parameter is being passed to the loader
in the TaskioRootGroup.__init__ making possible to create configs
with a custom root.

Fixes: #26
Fixes: #27